Add numba kernel for metropolis_exchange
