+++ netket/sampler/__init__.py
@@ -1,3 +1,5 @@
 from .abstract_sampler import AbstractSampler
 from .metropolis_hastings import *
+from .metropolis_local import *
+from .metropolis_exchange import *
 from .exact_sampler import *
